% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghmd.R
\name{gghmd}
\alias{gghmd}
\title{gghmd function}
\usage{
gghmd(my.df, loc = "USA")
}
\arguments{
\item{Take}{hmd_pop as input}
}
\value{
ggplot2 graph
}
\description{
This function loads a hmd_pop as input dataframe.
ggplot function is used here and we will get a simple country plot with available timeframe.
}
\examples{
gghmd(hmd_pop)
}
\keyword{"USA")}
\keyword{(Ex:}
\keyword{Country}
\keyword{as}
\keyword{code}
\keyword{loc}

