% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghmd_bar.R
\name{gghmd_bar}
\alias{gghmd_bar}
\title{gghmd_bar function}
\usage{
gghmd_bar(my.df, loc = "AUS", aageMin = 80, aageMax = 90, yrMin = 1980,
  yrMax = 2010)
}
\arguments{
\item{Take}{hmd_pop as input}
}
\value{
ggplot2 graph
}
\description{
contains the Year and the subsequent columns are the sample identifiers.
yrMin and yrMax can be used as time duration for the plot.
}
\examples{
#Example.1; base graph with default values ( yrMin=1960, yrMax=2010)
gghmd_bar(hmd_pop)
#Example.2; Advanced graph with user defined values
gghmd_bar(hmd_pop,"USA",80,90,1970,2000)

}
\keyword{,}
\keyword{loc}
\keyword{yrMax}
\keyword{yrMin,}

