% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghmd_loc.R
\name{gghmd_loc}
\alias{gghmd_loc}
\title{gghmd_loc function}
\usage{
gghmd_loc(my.df, loc = "USA", yrMin = 1960, yrMax = 2010)
}
\arguments{
\item{Take}{hmd_pop as input}
}
\value{
ggplot2 graph
}
\description{
This function loads a hmd_pop as input dataframe . It assumes that the first column
contains the Year and the subsequent columns are the sample identifiers.
yrMin and yrMax can be used as time duration for the plot.
}
\examples{
#Example.1; base graph with default values (loc="USA", yrMin=1980, yrMax=2010)
gghmd_loc(hmd_pop)
#Example.2; Advanced graph with user defined values
gghmd_loc(hmd_pop,"BEL",1970,2000)

}
\keyword{,}
\keyword{loc}
\keyword{yrMax}
\keyword{yrMin,}

