% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghmd_map.R
\name{gghmd_map}
\alias{gghmd_map}
\title{gghmd_map function}
\usage{
gghmd_map(df, yrMin = 1960, yrMax = 2010, yrDiv = 10, aage = 20)
}
\arguments{
\item{Take}{hmd_pop as input}
}
\value{
rworldmap graph
}
\description{
This function loads a hmd_pop as input dataframe . It assumes that the first column
contains the Year and the subsequent columns are the sample identifiers.
yrMin and yrMax can be used as time duration for the plot.
}
\examples{
#Example.1; base graph with default values (yrMin=1980, yrMax=2010,yrDiv,aage)
gghmd_map(hmd_pop)
#Example.2; Advanced graph with user defined values
gghmd_map(hmd_pop,1980,2010,10,30)

}
\keyword{yrMax,yrDiv,aage}
\keyword{yrMin,}

